# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Repair Shop Database"
  description: "Mock mechanic shop db"
  version: "1.2.0"
schemes:
- "http"
host: "127.0.0.1:5000"
consumes:
  - "application/json"
produces:
  - "application/json"
paths:
...
  /mechanics/login:
    post:
      tags:
        - Mechanics
      summary: "Login endpoint, (generates token)"
      description: "Validates mechanic credentials, generates and return a mechanic specific token. *Token times out in 1 hour*"
      parameters:
        - in: "body"
          name: "body"
          description: "Login Credentials"
          required: true
          schema:
            example:
              {
                "email": "ed@repairshop.com",
                "password": "test1"
              }
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
            example:
              application/json:
                {
                  "email": "ed@repairshop.com",
                  "password": "test1"
                }
        400:
          description: "Required data missing from body"
          schema: 
            $ref: "#/definitions/LoginMessage"
            example:
              application/json:
                {
                  "message": "Username and password required."
                }
        401:
          description: "Email or Password not found to match OR user.soft_delete=1 (check user 2 after delete)"
          schema:
            $ref: "#/definitions/LoginError"
            example:
              application/json:
                {
                  "error": "Invalid email or password!"
                }
  /mechanics/:
    post:
      tags:
        - Mechanics
      summary: "Create Mechanic"
      description: "Data object requires: name, email (*must be unique*), phone number (*11 character limit*) and password.\n*Limited to 5 per day*"
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: "#/definitions/CreateMechanicPayload"
      responses:
        201:
          description: "Mechanic creation successful"
          schema:
            $ref: "#/definitions/CreateMechanicResponse"
        400:
          description: "Required data missing from body"
          schema: 
            $ref: "#/definitions/Customer400Message"
            example:
              application/json:
                {
                  "message": "all data fields required"
                }
        409:
          description: "Email already associated with account - same email as first example. \n*Note* Testing duplicate email will reduce number of attempts per limiter"
          schema:
            $ref: "#/definitions/LoginError"
            example:
              application/json:
                {
                  "error": "'error details' integrity error."
                }
    
    get:
      tags:
        - Mechanics
        - Method Protection → Mechanic Token
      summary: "Get All Mechanics - PROTECTED METHOD - add 'Bearer ' and paste in the mechanic token"
      description: "Mechanics/Admin only\n *Example* /mechanics/?page=1&per_page=5"

      security:
        - bearerAuth: []

      parameters:
        - in: query
          name: page
          type: integer
          required: false
        - in: query
          name: per_page
          type: integer
          required: false

      responses:
        200:
          description: "All mechanics view for mechanics, pagination is optional, "
          schema:
            example:
              [
                {
                    "email": "ed@repairshop.com",
                    "id": 1,
                    "name": "ed",
                    "password": "test1",
                    "phone": "9719719711",
                    "salary": 199999.0
                },
                {
                    "email": "edd@repairshop.com",
                    "id": 2,
                    "name": "edd",
                    "password": "test2",
                    "phone": "5035035033",
                    "salary": 2.0
                }
              ]
  /mechanics/{mechanic_id}:
    put:
      tags:
      - Mechanics
      - Method Protection → Mechanic Token
      summary: "Change mechanic data - PROTECTED METHOD - add 'Bearer ' and paste in the token"
      description: "Mechanic id based on bearer token decode payload, Logged-in mechanic can change their data.\n*Change some data to test - Limit 6 per day*"
      
      security:
        - bearerAuth: []
      
      parameters:
        - in: path
          name: mechanic_id
          type: integer
          required: true
          description: The id of the mechanic who's data is changing
        - in: "body"
          name: "body"
          description: "The new data for the mechanic"
          required: true
          schema:
            $ref: ""
            example:
              {
                "name": "Ed",
                "email": "",
                "password": "",
                "phone": "3434433434",
                "salary": 0
              }
      
      responses:
        200:
          description: "Change successful"
          schema:
            $ref: ""
            description: "Returns all mechanic data\n*Changes in example:* Uppercase name field, and updated phone number"
            example:
              {
                "id": 1,
                "name": "Ed",
                "email": "ed@repairshop.com",
                "password": "test1",
                "phone": "3434433434",
                "salary": 290000
              }

        422:
          description: "All fields sent without data - no change"
          schema:
            example:
              {
                "message": "No changes made"
              }
        
        400:
          description: "Validation error - if no phone data is present in request"
          schema:
            $ref: "#/definitions/Customer400Message"
            example:
              application/json:
                {
                  "message": "{'phone': ['Missing data for required field.']} - all mechanic data fields required."
                }
    delete:
      tags:
      - Mechanics
      - Method Protection → Mechanic Token
      summary: "Soft delete mechanics - PROTECTED METHOD - add 'Bearer ' and paste in the token"
      description: "Set soft_delete parameter to 1 (true) - limit: 5 per day\n*NOTE: soft delete used to keep records for tax year or longer if recallable parts used in repair.*"
      
      security:
        - bearerAuth: []

      parameters:
        - in: path
          name: mechanic_id
          type: integer
          required: true

      responses:
        200:
          description: "Mechanic successfully marked for deletion"
          schema:
            example:
              {
                "message": "Mechanic successfully marked for deletion"
              }
        404:
          description: "Mechanic not found"
          schema:
            example:
              {
                "error": "Mechanic not found."
              }
  /mechanics/top-mechanics:
    get:
      tags:
      - Mechanics
      summary: "Get Mechanics by rank (number of assigned tickets)"
      description: "Ranked object with mechanic's name and the number of tickets to which they are assigned"

      parameters:
        - in: query
          name: limit
          type: integer
          description: "default value is 3"
          required: false
      
      responses:
        200:
          description: "Successful request"
          schema:
            example:
              {
                "1": {
                    "name": "ed",
                    "ticket_count": 4
                },
                "2": {
                    "name": "eddy",
                    "ticket_count": 3
                },
                "3": {
                    "name": "edd",
                    "ticket_count": 1
                }
              }
  /mechanics/current-customer-search:
    get:
      tags:
        - Mechanics
        - Method Protection → Mechanic Token
      summary: "Mechanics can search non-deleted customers by any piece of information"
      description: "Query terms: id, name, email, phone, or any"

      security:
        - bearerAuth: []

      parameters:
        - in: query
          name: id
          type: integer
          required: false
        - in: query
          name: name
          type: string
          required: false
        - in: query
          name: email
          type: string
          required: false
        - in: query
          name: phone
          type: string
          required: false
        - in: query
          name: any
          type: string
          required: false

      responses:
        200:
          description: "Successful request"
          schema:
            example:
              [
                {
                    "email": "kyle@gmail.com",
                    "id": 1,
                    "name": "kyle",
                    "phone": "9714175138",
                    "soft_delete": false
                },
                {
                    "email": "ace@gmail.com",
                    "id": 3,
                    "name": "aeslin",
                    "phone": "9713403030",
                    "soft_delete": false
                },
                {
                    "email": "dummy@gmail.com",
                    "id": 4,
                    "name": "dummy",
                    "phone": "5033286403",
                    "soft_delete": false
                },
                {
                    "email": "jaime@gmail.com",
                    "id": 5,
                    "name": "jaime",
                    "phone": "1234567890",
                    "soft_delete": false
                }
              ]

        404:
          description: "No results"
          examples:
            {
              "message": Filters failed to yield results.
            }
  /mechanics/deleted-customer-search:
    get:
      tags:
        - Mechanics
        - Method Protection → Mechanic Token
      summary: "Mechanics can search non-deleted customers by any piece of information"
      description: "Query terms: id, name, email, phone, or any"

      security:
        - bearerAuth: []

      parameters:
        - in: query
          name: id
          type: integer
          required: false
        - in: query
          name: name
          type: string
          required: false
        - in: query
          name: email
          type: string
          required: false
        - in: query
          name: phone
          type: string
          required: false
        - in: query
          name: any
          type: string
          required: false

      responses:
        200:
          description: "Successful request"
          schema:
            example:
              [
                {
                    "email": "revel@gmail.com",
                    "id": 2,
                    "name": "rev",
                    "phone": "5033286403",
                    "soft_delete": true
                }
              ]

        404:
          description: "No results"
          examples:
            {
              "message": Filters failed to yield results.
            }

# =================== SERVICE TICKETS =========================
  # ============== Customer Token Protected ==============
...
definitions: 
...
  LoginResponse:
    type: object
    properties:
      auth_token:
        type: string
      message:
        type: string
      status:
        type: string

    example:
      auth_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3NTE4MzE0ODEsImlhdCI6MTc1MTgyNzg4MSwic3ViIjoiMSIsInJvbGUiOiJjdXN0b21lciJ9.pCT3Ksn3aVgg8hEChUW1Z9hfDrRXiwnUThXK3gjD_fM"
      message: "Successfully Logged In"
      status: "success"
  LoginMessage:
    type: object
    properties:
      message:
        type: string
    example:
      message: "Username and password required."
  LoginError:
    type: object
    properties:
      error:
        type: string
    example:
      error: "Invalid email or password!"
...
  CreateMechanicPayload:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
      phone:
        type: string
        maxLength: 11
      salary:
        type: number
        format: float
    example:
      name: "ed"
      email: "ed@repairshop.com"
      password: "test1"
      phone: "0123456789"
      salary: 290000.00
...
  CreateMechanicResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      salary:
        type: number
        format: float
    example:
      id: 1
      name: "Ed"
      email: "ed@repairshop.com"
      password: "test1"
      phone: "0123456789"
      salary: 290000.00
...
  Customer400Message:
    type: object
    properties: 
      message:
        type: string
    example:
      message: "'error details' - all customer data fields required."
...
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header